requirements.txt
anyio==1.4.0
async-generator==1.10
asyncclick==7.0.9
certifi==2020.6.20
chardet==3.0.4
idna==2.10
importlib-metadata==1.7.0
influxdb==5.3.0
msgpack==0.6.1
python-dateutil==2.8.1
python-kasa==0.4.0.dev1
pytz==2020.1
requests==2.24.0
six==1.15.0
sniffio==1.1.0
urllib3==1.25.10
zipp==3.1.0